tools/watson-orchestrate/requirements.txt
cairosvg
jq
pillow
